add page and contollers for mechanic add orders

+++ src/main/java/com/car/service/impl/UsersServiceImpl.java
@@ -8,6 +8,7 @@ import com.car.entity.Orders;
 import com.car.entity.Roles;
 import com.car.entity.Users;
 import com.car.service.exception.ObjectAlreadyExistsException;
+import com.car.service.exception.UserNotFound;
 import com.car.service.interfaces.RolesService;
 import com.car.service.interfaces.UsersService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -51,51 +52,41 @@ public class UsersServiceImpl implements UsersService
     @Override
     public Users addUser(Users users, String secretCode) throws ObjectAlreadyExistsException
     {
-        if(users == null){
-            throw new NullPointerException();
+        Roles role;
+        Users userAdd;
+        if (userExist(users.getUsername())) {
+            throw new ObjectAlreadyExistsException("User with login " + users.getUsername() + " already exist");
+        } else {
+            if (secretCode.equals("admin")) {
+                role = rolesDAO.findByRole("ROLE_MECHANIC");
+            } else {
+                role = rolesDAO.findByRole("ROLE_USER");
+            }
+            List<Roles> roles = new ArrayList<>();
+            roles.add(role);
+            users.setRoleList(roles);
+            userAdd = usersDAO.save(users);
         }
-            Users userAdd = null;
-            List<Roles> role = new ArrayList<>();
-            Roles r = null;
-
-            Users findUser = factoryDTO.UsersOutDTO(usersDAO.findByUsername(users.getUsername()));
-           /* Users findUser = factoryDTO.UsersOutDTO(findUserById(users.getUserId()));*/
-
-            if (users.getUsername().equals(findUser.getUsername())) {
-                throw new ObjectAlreadyExistsException();
-            }
-            else
-            {
-                if (secretCode.equals("admin")) {
-                    r = rolesDAO.findOne(2L);
-                    role.add(r);
-                    users.setRoleList(role);
-                    userAdd = usersDAO.save(users);
-
-                } else {
-                    r = rolesDAO.findOne(1L);
-                    role.add(r);
-                    users.setRoleList(role);
-                    userAdd = usersDAO.save(users);
-                }
-            }
-            return factoryDTO.UsersOutDTO(userAdd);
+        return factoryDTO.UsersOutDTO(userAdd);
     }
 
     @Transactional
     @Override
-    public void deleteUser(Long userId)
+    public void deleteUser(Long userId) throws UserNotFound
     {
         Users deleteUser = findUserById(userId);
-        usersDAO.delete(deleteUser);
+        if (deleteUser == null) {
+            throw new UserNotFound();
+        } else {
+            usersDAO.delete(deleteUser);
+        }
     }
 
     @Transactional (readOnly = true)
     @Override
     public List<Orders> getUserOrders(Long userId)
     {
-        /*return usersDAO.getUserOrders(userId);*/
-                return em.createQuery("SELECT o.orderId, o.breaking, o.status FROM orders o WHERE o.userId = :userId", Orders.class)
+        return em.createQuery("SELECT o.orderId, o.breaking, o.status FROM orders o WHERE o.userId = :userId", Orders.class)
                 .setParameter("userId", 1)
                 .getResultList();
     }
@@ -104,7 +95,6 @@ public class UsersServiceImpl implements UsersService
     @Override
     public List<Car> getUserCars(Long userId)
     {
-        /*return usersDAO.getUserCars(user_id);*/
         return em.createQuery("SELECT c.carId, c.mark, c.color, c.miles, c.vin FROM cars c WHERE c.userId = :userId", Car.class)
                 .setParameter("userId", 1)
                 .getResultList();
@@ -127,6 +117,10 @@ public class UsersServiceImpl implements UsersService
     public List<Users> getAllUsers()
     {
         return (List<Users>) usersDAO.findAll();
+    }
+
+    private boolean userExist(String username) {
+        return usersDAO.findByUsername(username) != null;
     }
 
 }
